Full file tree restructure
