+++ Forward-client/app/components/layout/footer.tsx
