+++ Forward-client/app/components/ui/password-input.tsx
@@ -6,7 +6,7 @@ import { EyeIcon, EyeOffIcon } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/utils";
 
 const PasswordInput = React.forwardRef<
   HTMLInputElement,
