+++ Forward-client/app/components/ui/popover.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as PopoverPrimitive from "@radix-ui/react-popover";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/utils";
 
 function Popover({
   ...props
