+++ Forward-client/app/components/ui/tabs.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as TabsPrimitive from "@radix-ui/react-tabs";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/utils";
 
 function Tabs({
   className,
