+++ Forward-client/app/features/account/slices/userSlice.ts
@@ -1,0 +1,18 @@
+import { type PayloadAction, createSlice } from "@reduxjs/toolkit";
+import type { User } from "@/features/account/types";
+
+const initialState: { user: User | null } = { user: null };
+
+export const userSlice = createSlice({
+  name: "user",
+  initialState,
+  reducers: {
+    setUser: (state, action: PayloadAction<User | null>) => {
+      state.user = action.payload;
+    },
+  },
+});
+
+export const { setUser } = userSlice.actions;
+
+export default userSlice.reducer;
