+++ Forward-client/app/routes.ts
@@ -6,14 +6,14 @@ import {
 } from "@react-router/dev/routes";
 
 export default [
-  layout("./layout.tsx", [
+  layout("routes/layout.tsx", [
     index("routes/home.tsx"),
     route("login", "routes/login.tsx"),
     route("register", "routes/register.tsx"),
-    layout("protected.tsx", [
-      route("dashboard", "routes/dashboard.tsx"),
-      route("account", "routes/account.tsx"),
-      route("lesson/:lessonId", "routes/lesson.tsx"),
+    layout("routes/protected/protected.tsx", [
+      route("dashboard", "routes/protected/dashboard.tsx"),
+      route("account", "routes/protected/account.tsx"),
+      route("lesson/:lessonId", "routes/protected/lesson.tsx"),
     ]),
   ]),
 ] satisfies RouteConfig;
