+++ Forward-client/app/routes/login.tsx
@@ -3,10 +3,9 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { PasswordInput } from "@/components/ui/password-input";
 import { useState } from "react";
-import { useAuth } from "@/lib/useAuth";
-import type { User } from "@/lib/redux/userSlice";
+import { useAuth } from "@/features/account/hooks";
 import { toast } from "sonner";
-import { apiFetch } from "@/lib/utils";
+import { apiFetch } from "@/utils/utils";
 
 export default function Login() {
   const [error, setError] = useState(null);
